Fix jobs post user (#121)
* Add user_id to postJob action

* Update handleClosePanel to setstate cb

* Cleanup and /update todo comment

* Rework postJobs to use user_id from req.user instead of redux

* Add user_id to jobs routing and queries

* Comment cleanup

+++ server/routes/jobs.js
@@ -4,8 +4,7 @@ const { requireAuthToken } = require("../middleware/helper");
 
 module.exports = app => {
   app.post("/jobs", requireAuthToken, async (req, res, next) => {
-    // TODO - Need to add user_id from client req (action creator)
-
+    const user_id = req.user.id;
     const {
       posting_url,
       company_name,
@@ -14,8 +13,7 @@ module.exports = app => {
       current_status,
       job_location,
       job_source,
-      active,
-      user_id
+      active
     } = req.body;
 
     const insert = `
@@ -51,6 +49,8 @@ module.exports = app => {
   });
 
   app.post("/jobs/update", requireAuthToken, async (req, res, next) => {
+    // TODO - This needs to be fixed to handle updates properly
+    const user_id = req.user.id;
     const {
       id,
       posting_url,
@@ -72,8 +72,12 @@ module.exports = app => {
         current_status = $6,
         job_location = $7,
         job_source = $8,
-        active = $9
-      WHERE id = $1`;
+        active = $9,
+        updated = now()::timestamp
+      WHERE id = $1 AND user_id = $10
+      RETURNING *
+      ;`;
+
     try {
       const dbResponse = await pool.query(insert, [
         id,
@@ -84,32 +88,35 @@ module.exports = app => {
         current_status,
         job_location,
         job_source,
-        active
+        active,
+        user_id
       ]);
-      return res.json(dbResponse);
+      console.log(dbResponse.rows[0]);
+      return res.send({ status: 200 });
     } catch (err) {
       next(err);
     }
   });
 
   app.post("/jobs/delete", requireAuthToken, async (req, res, next) => {
+    const user_id = req.user.id;
     const { id } = req.body;
-    const remove = `DELETE FROM jobs WHERE id = $1`;
+    const remove = `DELETE FROM jobs WHERE id = $1 AND user_id = $2;`;
 
     try {
-      const dbResponse = await pool.query(remove, [id]);
-      return res.json(dbResponse);
+      const dbResponse = await pool.query(remove, [id, user_id]);
+      return res.send({ status: 200 });
     } catch (err) {
       next(err);
     }
   });
 
   app.get("/jobs", requireAuthToken, async (req, res, next) => {
-    // TODO - need to add user_id to query based on auth middleware/ request
-    const query = `SELECT * from jobs`;
+    const user_id = req.user.id;
+    const query = `SELECT * FROM jobs WHERE user_id = $1;`;
     try {
       // TODO add limit, sort and pagination to complete endpoint
-      const jobs = await pool.query(query);
+      const jobs = await pool.query(query, [user_id]);
       return res.json(jobs.rows);
     } catch (err) {
       next(err);
